feat: add cart and profile buttons to header
