+++ src/widgets/header/index.ts
@@ -1,3 +1,3 @@
-import LayoutHeader from './ui/layoutHeader/LayoutHeader';
+import LayoutHeader from './ui/layoutHeader';
 
 export default LayoutHeader;
