--- src/widgets/header/ui/layoutHeader/LayoutHeader.tsx
@@ -1,5 +1,0 @@
-function LayoutHeader(): JSX.Element {
-  return <div>Header</div>;
-}
-
-export default LayoutHeader;
