--- src/widgets/header/ui/nav/cart/HeaderNavCart.module.scss
@@ -1,3 +1,0 @@
-.header__nav_cart {
-  font-size: 16px;
-}
