--- src/widgets/header/ui/nav/cart/index.tsx
@@ -1,7 +1,0 @@
-import styles from './HeaderNavCart.module.scss';
-
-function HeaderNavCart(): JSX.Element {
-  return <div className={styles.header__nav_cart}>Header Nav Cart</div>;
-}
-
-export default HeaderNavCart;
