--- src/widgets/header/ui/nav/profile/HeaderNavProfile.module.scss
@@ -1,3 +1,0 @@
-.header__nav_profile {
-  font-size: 16px;
-}
