--- src/widgets/header/ui/nav/profile/index.tsx
@@ -1,7 +1,0 @@
-import styles from './HeaderNavProfile.module.scss';
-
-function HeaderNavProfile(): JSX.Element {
-  return <div className={styles.header__nav_profile}>Header Nav Profile</div>;
-}
-
-export default HeaderNavProfile;
